==> scripts/02_Landuse_Soiltype/04_GeoData.ipynb <==
# Description
___

This script writes the GeoData file from the sorted SLCs, shapefiles and elevation/coordinate agnostic outputs. Columns will not be properly sorted.

import os
import pandas as pd
import numpy as np
import geopandas as gpd

### Inputs

# Read the modified catchment shapefile
modifiedcat = gpd.read_file('../../geospacial/shapefiles/modified_shapefiles/Modified_SMMcat.shp')

# Read the river shapefile
riv = gpd.read_file('../../geospacial/shapefiles/river_shapefile/smm_riv.shp')

# Import latitude, longitude and adjusted elevation from gistool
lat_elev=pd.read_csv('../../geospacial/misc/elev_lat_stats_elv.csv')

slc = pd.read_csv('../../geospacial/slc/sorted_final_slc.csv')

# Open the text file in read mode
with open('../../geospacial/misc/downstream_order.txt', 'r') as file:
    # Read the lines of the file into a list
    lines = file.readlines()

# Optionally, you may strip newline characters from each line
order = [line.strip() for line in lines]

### Create Geodata

# Sort the GeoDataFrame by a specific column
sorted_modifiedcat = modifiedcat.sort_values(by='seg_nhm')
sorted_riv = riv.sort_values(by='seg_nhm')

# Select the added river ID rows based on the range in a column
selected_rows = modifiedcat[(modifiedcat['seg_nhm'] >= 58662 )]

# Concatenate the selected rows with the river shapefile
merged_data = pd.concat([sorted_riv, selected_rows], ignore_index=True)

# Convert the merged data to a GeoDataFrame
merged_riv = gpd.GeoDataFrame(merged_data, crs=riv.crs)

merged_riv=merged_riv.fillna(0)

Add columns

# Extract the first two columns from the merged_riv GeoDataFrame (subid and maindown)
geodata = merged_riv.iloc[:, :2]

# Add subbasin area from catchment shapefile based on river segment
geodata = pd.merge(geodata, modifiedcat[['seg_nhm', 'Shape_Area']], on='seg_nhm')

geodata

# Merge the lat lon elev data with modifiedcat based on hru
lat_elev= pd.merge(modifiedcat, lat_elev, left_on='hru_nhm', right_on='hru_nhm', how='inner')

# merge lat lon elev info into geodata based on river ID
geodata = pd.merge(geodata, lat_elev[['seg_nhm', 'lon', 'lat', 'mean']], on='seg_nhm')

# merge river slope to the geodata from merged_riv
geodata = pd.merge(geodata, merged_riv[['seg_nhm', 'seg_slope']], on='seg_nhm')

# merge river length to the geodata from merged_riv
geodata = pd.merge(geodata, merged_riv[['seg_nhm', 'Shape_Leng']], on='seg_nhm')

geodata

Add SLC data to geodata

slc

# Merge the data from the CSV file with modifiedcat
slc = pd.merge(modifiedcat, slc, left_on='hru_nhm', right_on='Unnamed: 0', how='inner')

slc

# Count the amount of slc columns
column_range = []
for column in slc.columns:
    # Check if the column header consists of either 3 or 4 digits
    if column.isdigit() and (len(column) == 3 or len(column) == 4):
        column_range.append(column)

len(column_range)

# Merge the selected columns from merged_data with data_to_write
final_geodata = pd.merge(geodata, slc[['seg_nhm'] + list(column_range)],  on='seg_nhm')
final_geodata

# set downstream values of 0 to -9999
final_geodata.loc[final_geodata['ds_seg_nhm'] == 0, 'ds_seg_nhm'] = -9999

final_geodata

#Rename the first columns to HYPE GeoData format
final_geodata.rename(columns={'seg_nhm':'subid' , 'ds_seg_nhm':'maindown', 'Shape_Area':'area', 'lon': 'longitude','lat': 'latitude', 'mean':'elev_mean', 'seg_slope':'slope_mean', 'Shape_Leng':'rivlen'}, inplace=True)   

Rename columns

# Generate the new column names
new_column_names = ['SLC_' + str(i + 1) for i in range(len(column_range))]

# Create a dictionary mapping old column names to new column names
column_name_mapping = {old_name: new_name for old_name, new_name in zip(column_range, new_column_names)}

# Rename the columns using the mapping
final_geodata.rename(columns=column_name_mapping, inplace=True)

final_geodata

Sort geodata

# Convert string values to integers
order = list(map(int, order))

# Set 'seg_nhm' column as the index
final_geodata.set_index('subid', inplace=True)

# Convert index values to integers
final_geodata.index = final_geodata.index.astype(int)

# Reindex the DataFrame based on the order specified in the 'order' list
final_geodata_sorted = final_geodata.reindex(order)

final_geodata_sorted

final_geodata_sorted.to_csv('../../model/add_HDS_GeoData.txt', sep='\t', index=True)
