==> scripts/02_Landuse_Soiltype/06_add_depression_depths.ipynb <==
# Description
___

This script adds HDS parametrization to GeoData extracted from NCAMapper.

import pandas as pd
import geopandas as gpd
import networkx as nx
import numpy as np

### Read Data

geodata = pd.read_csv('../../model/add_HDS_GeoData.txt', header=0, index_col=0, sep='\t')

hds_depths= gpd.read_file('../../geospacial/depressions/ncamapper_depressions.gpkg')

# define ilake slcs for the Milk
milk_ilake= 'SLC_62'

### Remove HDS from St. Mary

    # Creating a DiGraph out of `df` object
riv_graph = nx.from_pandas_edgelist(geodata.reset_index(), source='subid', target='maindown', create_using=nx.DiGraph)

# Find St. Mary Segments
stmary = [58183]
stmary.extend(nx.ancestors(riv_graph, 58183))

len(stmary)

# Remove all rows containing info on St Mary so only the Milk is 
hdsdepths_filtered = hds_depths[~hds_depths['seg_nhm'].isin(stmary)]

### Format GeoData

# Set 'seg_nhm' as the index in hdsdepths_filtered
hdsdepths_filtered = hdsdepths_filtered.set_index('seg_nhm')


# Merge based on the index
merged_geodata = geodata.merge(hdsdepths_filtered[['_count', '_mean']], left_index=True, right_index=True, how='left')

# Replace NaN values with 0 in the new columns
merged_geodata['_count'].fillna(0, inplace=True)
merged_geodata['_mean'].fillna(0, inplace=True)

# Rename '_mean' to 'hds_depth'
merged_geodata.rename(columns={'_mean': 'hds_depth'}, inplace=True)

# find index of first and last SLC
filtered_columns = merged_geodata.filter(like='SLC').columns

if len(filtered_columns) > 0:
    first_slc_index = merged_geodata.columns.get_loc(filtered_columns[0])
    last_slc_index = merged_geodata.columns.get_loc(filtered_columns[-1])
    print("First column index with 'SLC':", first_slc_index)
    print("Last column index with 'SLC':", last_slc_index)
else:
    print("No columns with 'SLC' in the name found.")

# Convert count to area
merged_geodata['_count'] = merged_geodata['_count'] * (30 * 30)

# Add a new column 'hds_frac' which is equal to count/area
merged_geodata['hds_frac'] = merged_geodata['_count'] / merged_geodata['area']

merged_geodata[milk_ilake] = merged_geodata['hds_frac']

### Adjust Milk SLC fractions based on weights

# Iterate through each row
for index, row in merged_geodata.iterrows():
    
        # Calculate the sum of slc columns for the current row
        sum_of_columns_row = row.iloc[first_slc_index:last_slc_index+1].sum()

        # Subtract the value in the milk_ilake column for the current row
        result_row = sum_of_columns_row - row[milk_ilake]

        # Calculate the scaling factor based on the formula
        scaling_factor = 1 - row[milk_ilake]

        # Iterate through SLC_1 to SLC_n columns and update values
        for col in merged_geodata.columns:
            if col.startswith('SLC_') and col != milk_ilake:
                merged_geodata.loc[index, col] = (row[col] / result_row) * scaling_factor


### Check that SLCs still sum to 1

# Check sums of rows
for index, row in merged_geodata.iterrows():
    # Calculate the sum of values in the specified columns
    row_sum = row.iloc[first_slc_index:(last_slc_index + 1)].sum()

    # Check if the sum is approximately equal to 1
    if not np.isclose(row_sum, 1, rtol=1e-6):
        print(f"Warning: Row {index} does not sum to 1 (Sum: {row_sum})")

    # Insert 'hds_depth' column just before the first 'SLC' column
    columns = list(merged_geodata.columns)
    columns.insert(first_slc_index, columns.pop(columns.index('hds_depth')))
    merged_geodata = merged_geodata[columns]

# Delete the 'hds_frac' column
merged_geodata.drop(columns=['hds_frac'], inplace=True)
# Delete the '_count' column
merged_geodata.drop(columns=['_count'], inplace=True)

# Save the GeoDataFrame as a tab-separated text file
merged_geodata.to_csv('../../model/GeoData.txt', sep='\t', index=True)
